refactor(core): extract relative-time formatter as an Instant extension
Move the private "x ago" formatter out of UserListItem into a reusable
Instant.formatRelativeTime() in core/presentation (returns null for the
epoch-zero sentinel), and have UserListItem use it. No behaviour change.
